# routes/users.js
const router = require("express").Router();
const User = require("../models/User");
const Post = require("../models/Post");
const bcrypt = require("bcrypt");
const { Mongoose } = require("mongoose");

//UPDATE User

router.put("/:id",async (req,res)=>{
    if(req.body.userId === req.params.id){
        if(req.body.password){
            const salt = await bcrypt.genSalt(10);
            req.body.password = await bcrypt.hash(req.body.password,salt)
        }
        try{
            const updateUser = await  User.findByIdAndUpdate(req.params.id,{
                $set: req.body,
            },{new:true});
            res.status(200).json(updateUser);
        } catch(err){
            res.status(500).json(err);
        }
    }
    else{
        res.status(401).json("You can update only your accout!");
    }
});

router.put("/sp/:id",async (req,res)=>{
    if(true){
        if(req.body.password){
            const salt = await bcrypt.genSalt(10);
            req.body.password = await bcrypt.hash(req.body.password,salt)
        }
        try{
            const updateUser = await  User.findByIdAndUpdate(req.params.id,{
                $set: req.body,
            },{new:true});
            res.status(200).json(updateUser);
        } catch(err){
            res.status(500).json(err);
        }
    }
    else{
        res.status(401).json("You can update only your accout!");
    }
});

 
//DELETE User

router.delete("/:id",async (req,res)=>{
    if(req.body.userId === req.params.id){
        try{
            const user = await User.findById(req.params.id);
            try{
                await Post.deleteMany({username:user.username})
                await User.findByIdAndDelete(req.params.id)
                res.status(200).json("User has been deleted ...");
            } catch(err){
                res.status(500).json(err);
            }
        }catch(err){
            res.status(404).json("User not found!");
        }
    }
    else{
        res.status(401).json("You can delete only your accout!");
    }
    
});

router.delete("/sp/:id",async (req,res)=>{
    if(true){
        try{
            const user = await User.findById(req.params.id);
            try{
                await Post.deleteMany({username:user.username})
                await User.findByIdAndDelete(req.params.id)
                res.status(200).json("User has been deleted ...");
            } catch(err){
                res.status(500).json(err);
            }
        }catch(err){
            res.status(404).json("User not found!");
        }
    }
    else{
        res.status(401).json("You can delete only your accout!");
    }
    
});


//GET USER

router.get("/:id", async (req,res)=>{
    try{
        const user = await User.findById(req.params.id);
        const {password, ...others} = user._doc;
        res.status(200).json(others);
    }catch(err){
        res.status(500).json(err)
    }
})

// Get all users
router.get("/",async (req,res)=>{
    try {
        const users = await User.find();
        res.status(200).json(users);
        
    } catch (err) {
        res.status(500).json(err);
    }

})

module.exports = router